requirements.txt
firecrawl==1.12.0
supabase==1.0.3
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
apscheduler==3.10.4
pydantic>=2.10.3
httpx>=0.23.0,<0.24.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
aiosqlite==0.19.0
asyncpg==0.29.0
jinja2==3.1.3
beautifulsoup4==4.12.3
lxml==5.1.0  # For faster HTML parsing
httpx>=0.23.0,<0.24.0  # Compatible version 